refactor: argparsed moved to separated class
