+++ .gitignore
@@ -1,7 +1,5 @@
 .vscode/
-/src/dictionary/__pycache__
-/src/anki/__pycache__
-/src/dictionary/languages/__pycache__
+__pycache__
 venv
 build
 dist
